Fixes #23: Moved Helper class to parent class package to access protected methods without inheritance.
